+++ src/router/routes.js
@@ -4,9 +4,9 @@ const routes = [
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
     children: [
-      { path: '', component: () => import('src/pages/PDashboard.vue') },
-      { path: 'ibindex', component: () => import('src/pages/PIbindex.vue') },
-      { path: 'ibindex-weights', component: () => import('src/pages/PIbindexWeights.vue') },
+      { path: '', component: () => import('src/pages/PageDashboard.vue') },
+      { path: 'ibindex', component: () => import('src/pages/PageIbindex.vue') },
+      { path: 'ibindex-weights', component: () => import('src/pages/PageIbindexWeights.vue') },
       { path: 'funds', component: () => import('src/pages/PageFunds.vue') }
     ]
   },